pyproject.toml
[project]
name = "things-mcp"
version = "0.1.1"
description = "Things app Model Context Protocol (MCP) server for Claude and Anthropic Windsurf"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT License"}
authors = [
    {name = "Harald Lindstrøm", email = "hald@users.noreply.github.com"}
]
keywords = ["mcp", "anthropic", "claude", "things", "task-management", "productivity"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
    "Topic :: Office/Business :: Scheduling",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: MCP"
]
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.2.0",
    "things-py>=0.0.15",
]

[project.scripts]
things-mcp = "src.things_mcp.fast_server:run_things_mcp_server"

[project.urls]
"Homepage" = "https://github.com/hald/things-mcp"
"Bug Tracker" = "https://github.com/hald/things-mcp/issues"
"Documentation" = "https://github.com/hald/things-mcp#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/things_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "src/things_mcp",
    "README.md",
    "LICENSE",
    "pyproject.toml",
    "things_fast_server.py",
    "things_server.py"
]

[tool.hatch.envs.default]
dependencies = [
    "pytest",
    "pytest-cov",
    "ruff",
    "build",
    "twine"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "build>=0.10.0",
    "twine>=4.0.0"
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
lint = "ruff check ."
format = "ruff format ."
build = "python -m build"
publish = "twine upload dist/*"
clean = "rm -rf dist build *.egg-info"
